feat(modbus): 优化 Modbus TCP 协议支持并添加设备配置
- 更新 DeviceConfig 结构，将 SlaveID 类型从 byte 改为 int
- 修改 devices.yaml，更新设备配置示例- 重构 main.go，优化设备元数据注入和客户端池化- 更新 modbus_tcp.go，添加 SetSlave 方法动态设置从站 ID
- 在 points.yaml 中添加新设备的点位配置示例

+++ types/device.go
@@ -10,7 +10,7 @@ type DeviceMeta struct {
 	EnablePing   bool   `yaml:"enable_ping"`
 	IP           string `yaml:"ip"`
 	Port         int    `yaml:"port"`
-	SlaveID      byte   `yaml:"slave_id"` // 新增字段
+	SlaveID      int    `yaml:"slave_id"` // 新增字段
 }
 
 type DeviceConfigWithMeta struct {
